Create RecognizeRegularSingularMahlerSystem.py
This program has been written by M. Poulet and C. Faverjon. It takes as an input a Mahler system - that is an invertible matrix with rational function entries and an integer p>=2 - and it returns whether the system is regular singular at 0 or not. When it is the case, it returns an equivalent constant matrix and the first coefficients of the Puiseux development of an associated gauge transformation.